pyproject.toml
[project]
name = "openwhale"
version = "0.1.0"
description = "渗透测试智能体 - OpenWhale"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.94.0",
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "mcp[cli]>=1.27.0",
    "openai>=1.109.1",
    "python-dotenv>=1.2.2",
    "rich>=15.0.0",
    "uvicorn[standard]>=0.44.0",
]

[project.scripts]
openwhale = "openwhale.main:main"
openwhale-web = "openwhale.web.app:run"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/openwhale"]
